asdf

+++ swing-stocks-wei/src/portfolio.py
@@ -172,7 +172,7 @@ def run_portfolio(days: list[date], symbols: list[str], categories: dict[str, st
                     entry_price=float(price), exit_price=None,
                     gross_return_pct=None, holding_days=None,
                     target_weight_pct=weight * 100.0,
-                    effective_weight_pct=budget / equity_now * 100.0 if equity_now > 0 else 0.0,
+                    effective_weight_pct=float(budget / equity_now * 100.0) if equity_now > 0 else 0.0,
                     tier=tier,
                     cat_mom_at_entry=None if np.isnan(mom[s]) else mom[s],
                     is_open=False,
